# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Daftar Properti API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/photo/{fileId}/{fileName}": {
            "get": {
                "tags": [
                    "Image"
                ],
                "summary": "Show image",
                "operationId": "image.show",
                "parameters": [
                    {
                        "name": "fileId",
                        "in": "path",
                        "description": "File Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "fileName",
                        "in": "path",
                        "description": "Filename",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "image/*": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tele-app/upload/image": {
            "post": {
                "tags": [
                    "Image"
                ],
                "summary": "Upload Image",
                "operationId": "image.upload",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ImageUploadRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "fileId": {
                                            "type": "integer",
                                            "example": 123
                                        },
                                        "fileName": {
                                            "type": "string",
                                            "example": "image.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ImageUploadRequest": {
                "properties": {
                    "image": {
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
...
        }
    }
}